Add support for generic integer type conversions to Expression::Constant (#333)
Current implementation only provides `From<usize>`, which requires
explicit type conversions in various locations. This PR provides support
for type conversions for arbitrary primitive integer types,
specifically: `u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128,
isize`

In reference to [this
comment](https://github.com/scroll-tech/ceno/pull/305#discussion_r1789316123).

---------

+++ ceno_zkvm/src/instructions/riscv/b_insn.rs
@@ -62,9 +62,9 @@ impl<E: ExtensionField> BInstructionConfig<E> {
         // Fetch instruction
         circuit_builder.lk_fetch(&InsnRecord::new(
             vm_state.pc.expr(),
-            (insn_kind.codes().opcode as usize).into(),
+            insn_kind.codes().opcode.into(),
             0.into(),
-            (insn_kind.codes().func3 as usize).into(),
+            insn_kind.codes().func3.into(),
             rs1.id.expr(),
             rs2.id.expr(),
             imm.expr(),
