+++ ceno_zkvm/src/instructions/riscv/i_insn.rs
@@ -44,9 +44,9 @@ impl<E: ExtensionField> IInstructionConfig<E> {
         // Fetch the instruction.
         circuit_builder.lk_fetch(&InsnRecord::new(
             vm_state.pc.expr(),
-            (insn_kind.codes().opcode as usize).into(),
+            insn_kind.codes().opcode.into(),
             rd.id.expr(),
-            (insn_kind.codes().func3 as usize).into(),
+            insn_kind.codes().func3.into(),
             rs1.id.expr(),
             0.into(),
             imm.clone(),
